+++ test/Sentry.Tests/Protocol/Context/ContextsTests.cs
@@ -100,6 +100,26 @@ public class ContextsTests
     }
 
     [Fact]
+    public void SerializeObject_SingleResponsePropertySet_SerializeSingleProperty()
+    {
+        var sut = new Contexts
+        {
+            Response =
+            {
+                StatusCode = 200
+            }
+        };
+
+        var actualString = sut.ToJsonString(_testOutputLogger);
+
+        var actual = Json.Parse(actualString, Contexts.FromJson);
+        actual.Should().BeEquivalentTo(sut);
+
+        var expectedString = "{\"response\":{\"type\":\"response\",\"status_code\":200}}";
+        Assert.Equal(expectedString, actualString);
+    }
+
+    [Fact]
     public void SerializeObject_SingleRuntimePropertySet_SerializeSingleProperty()
     {
         var sut = new Contexts